+++ R/Smisc-package.R
@@ -1,13 +1,13 @@
-##' Sego Miscellaneous
+##' Smisc: Sego Miscellaneous
 ##'
-##' Sego Miscellaneous:  A collection of functions for statistical computing and data manipulation
+##' A collection of functions for statistical computing and data manipulation
 ##'
 ##' \tabular{ll}{
 ##' Package: \tab Smisc\cr
 ##' Type: \tab Package\cr
 ##' Version: \tab 0.3.2\cr
 ##' Date: \tab 2016-04-13\cr
-##' License: \tab LICENSE\cr
+##' License: \tab file LICENSE\cr
 ##' }
 ##'
 ##' @name Smisc-package